Revise IR emitter so 1 place of AllocaInst would be emitted in the entry BB
It seems AMDGPU backend in LLVM 5.0.0 release has a bug that alloca in BB other
than the entry one would has trouble be lowered. Most alloca would be
eliminated by SROA but not the particular one modified in this commit.

Use EmitAllocaAtFunctionEntry to get around the issue in AMDGPU backend.