Change to add "-DEIGEN_USE_HIP" as a default CXXOPT
This checkin actually does two things.

1. Add "-DEIGEN_USE_HIP" as a default CXXOPT in the Tensorflow build
2. Remove "#define EIGEN_USE_HIP" from all files

The macro EIGEN_USE_HIP is used to communicate to Eigen to use HIP header files (instead of the CUDA header files that i uses by default). It needs to defined for all Tensorflow kernel implementations (CPU and GPU). Prior to this change it was getting defined only in files where the EIGEN_USE_GPU macro was also defined.