[ROCm] StreamExecutor logic for ROCm platform
The commit contains StreamExecutor logic for ROCm platform. It includes
integration logic with major components on ROCm platform:

- HIP runtime APIs
- rocBLAS for BLAS routines
- MIOpen for convolution / pooling / batch normalization and other accelerated
  algorithms for DNN
- rocFFT for FFT routines
- rocRAND for RNG

Also included are relevant changes to:

- bazel script to build ROCm StreamExecutor on ROCm platform
- some StreamExecutor interfaces and CUDA implementation
- GPU common runtime related to interface changes in StreamExecutor
- operators affected by interface changes in StreamExecutor

Interface changes:

1) StreamInterface::CudaStreamMemberHack()

Despite the fact that StreamExecutor and GPU common runtime are largely
orthogonal, this particular routine in StreamExecutor is used in GPU common
runtime and a couple of other operators. In this commit it's renamed as
StreamInterface::GPUStreamMemberHack() and their call sites are also changed.

2) StreamExecutorInterface::CudaContextHack()

This member is renamed to StramExecutorInterface::GPUContextHack().

3) Forward and backward pooling interfaces in dnn.h

In order to support MIOpen, the functional equivalent library for CuDNN on
ROCm platform, a pointer to an instance of ScratchAllocator has to be added to
corresponding interfaces and implementations on CUDA and ROCm.

4) AlgorithmConfig in dnn.h

Scratch size is added in AlgorithmConfig to support algorithm finding in
MIOpen.

Authors:

- Jack Chung: jack.chung@amd.com
- Deven Desai: deven.desai@amd.com
- Johannes M Dieterich: Johannes.Dieterich@amd.com
- Peng Sun: Peng.Sun@amd.com
- Jeffrey Poznanovic: Jeffrey.Poznanovic@amd.com