Enhancing the AlgorithmDesc data-structure for ROCm requirements
Currently the AlgoritmDesc data-structure has 2 fields
  * algo_id,
  * use_tensor_ops

The (algo_id, use_tensor_ops) tuple is sufficient to uniquely describe an algorithm on the CUDA platform. However on the ROCm, we instead need a (algo_id, solver, use_tensor_ops) tuple to unqiuely describe an algorihtm. On the ROCm platform, "algo_id" really refers to a family of "solvers" and hence an additional "solver" field is needed to uniquely describe a algorithm implementation.

Each unique algorithm implementaion also has a workspace size requirement associated with it. So adding it as field on the AlgorithmDesc data-structure makes sense. This will eliminate the need for the TF application code to query the CUDA/ROCm API to retrieve the worspace size everytime the convolution routines are invoked.

Enhancing the AlgorithmDesc data-structure to contain this extra-fields will allow us to maintain a common stream_executor/dnn interface for the CUDA and ROCm platforms. The other alternative would have been to modify that stream_executor/dnn interface to store/pass the ROCm specific fields/info.