Adding scratch_size field to dnn::AlgorithmDesc
Currently the scratch_size (workspace memory size), required to execute
a given convolution using a given algorithm, is determined via a call to
the CUDNN API. This call happens each time a convolution kernel is
called, because the scratch_size is not stored on neither the
dnn::AlgorithmDesc nor the dnn::AlgorithmConfig datastructures. Adding
scratch_size to dnn::AlgorithmDesc will remove the need to call the
CUDNN API to retrieve this information everytime we need to do
convolution.

There are also some ROCm platform specific reasons for making this change.

1. on the ROCm platform, there is no equivalent of the CUDNN API to
   determine the workspace memory size corresponding to a given
   (conv_params, algo) tuple. There is a ROCm API to determine the
   workspace memory size for a given conv_params, but that gives a
   worst case answer since that API does not take a specific algo as an
   input. As a consequence, on the ROCm fork of TF, we maintain a
   scratch_size field on the dnn::AlgorithmConfig data-structure. This
   change will remove the need for this discrepancy in the ROCm TF fork.

2. A side-effect of this commit is that the GetConvolve*Algorithms
   routines will now take the conv_params (input, filter, conv, output)
   as input. This will allow TF code to leverage underlying CUDNN /
   ROCm MIOpen APIs that return a list of algos that are applicable to
   the given conv_params. Currently the list of algorithms list is
   compiled in the TF code itself, without querying the CUDNN / MIOpen
   APIs.